+++ Exodus/src/utopia/exodus/rest/resource/user/MeNode.scala
@@ -3,10 +3,9 @@ package utopia.exodus.rest.resource.user
 import utopia.access.http.Method.Get
 import utopia.access.http.Status.Unauthorized
 import utopia.exodus.database.access.single.DbUser
-import utopia.exodus.rest.util.AuthorizedContext
+import utopia.exodus.rest.resource.scalable.{ExtendableSessionResource, SessionUseCaseImplementation}
 import utopia.flow.generic.ValueConversions._
-import utopia.nexus.http.Path
-import utopia.nexus.rest.ResourceWithChildren
+import utopia.nexus.rest.scalable.FollowImplementation
 import utopia.nexus.result.Result
 import utopia.vault.database.Connection
 
@@ -15,18 +14,11 @@ import utopia.vault.database.Connection
   * @author Mikko Hilpinen
   * @since 6.5.2020, v1
   */
-object MeNode extends ResourceWithChildren[AuthorizedContext]
+object MeNode extends ExtendableSessionResource
 {
 	override val name = "me"
 	
-	override val children = Vector(MyInvitationsNode, MyOrganizationsNode, MyLanguagesNode, MySettingsNode,
-		MyPasswordNode)
-	
-	override val allowedMethods = Vector(Get)
-	
-	override def toResponse(remainingPath: Option[Path])(implicit context: AuthorizedContext) =
-	{
-		context.sessionKeyAuthorized { (session, connection) =>
+	private val defaultGet = SessionUseCaseImplementation.default(Get) { (session, connection, _, _) =>
 			implicit val c: Connection = connection
 			// Reads user data and adds linked data
 			DbUser(session.userId).withLinks match
@@ -37,5 +29,9 @@ object MeNode extends ResourceWithChildren[AuthorizedContext]
 					Result.Failure(Unauthorized, "User no longer exists")
 			}
 		}
-	}
+	
+	override protected val defaultUseCaseImplementations = Vector(defaultGet)
+	override protected val defaultFollowImplementations =
+		Vector(MySettingsNode, MyOrganizationsNode, MyInvitationsNode, MyLanguagesNode, MyPasswordNode)
+			.map { FollowImplementation.withChild(_) }
 }